mcp/requirements.txt
mcp[cli]
weaviate-client==4.16.7
requests==2.32.4